Move Psr classes to separate namespace and rename classes
